+++ src/dialogs/ManageFilters.js
@@ -7,7 +7,7 @@ var popMenu = require('pop-menu');
 var automat = require('automat');
 
 var Dialog = require('./Dialog');
-var markup = require('../../html');
+var markup = require('../../assets/html');
 var copyInput = require('../lib/copy-input');
 
 var tabProperties = {
